bin/start_img.sh: better reboot check test

+++ bin/start_img.sh
@@ -25,8 +25,11 @@ fi
 sleep=0
 uptime --pretty | cut -f3 -d ' ' | grep -q "minutes"
 if [[ $? -eq 0 ]]; then
-  sleep=300
-  rm -f amd64-*/tmp/{LOCK,STOP} # cleanup
+  min=$(uptime --pretty | cut -f2 -d ' ')
+  if [[ $min -lt 2 ]]; then
+    sleep=120
+    rm -f amd64-*/tmp/{LOCK,STOP} # cleanup
+  fi
 fi
 
 for mnt in ${@:-~/amd64-*}
